check for existsing process from the second attempt

+++ deployment/bootstrap.sh
@@ -202,15 +202,15 @@ else
     sudo nginx
 fi
 
-
-
+echo "Stated nginx successfully"
 # Find the PID of any existing brainwars process
-PID=$(pgrep -f "./brainwars")
-
-if [ -n "$PID" ]; then
-    echo "brainwars is already running with PID $PID. Stopping..."
-    kill "$PID"
-    sleep 1  # give it a moment to shut down
+if [ -f "brainwars" ]; then
+    PID=$(pgrep -f "./brainwars")
+    if [ -n "$PID" ]; then
+        echo "brainwars is already running with PID $PID. Stopping..."
+        kill "$PID"
+        sleep 1  # give it a moment to shut down
+    fi
 fi
 
 # start the backend server
